Restructured Footer and Partners
