+++ src/components/footer/index.js
@@ -1,57 +1,10 @@
 import React from "react";
 import Link from "next/link";
-import styled from "styled-components";
-
-const Wrapper = styled.footer`
-  background-color: var(--lightBrown);
-  padding: 100px 100px 50px 100px;
-  color: var(--light);
-`;
-
-const FooterMenu = styled.div`
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
-  text-align: center;
-
-  h4 {
-    margin-bottom: 30px;
-    text-transform: uppercase;
-  }
-
-  ul {
-    margin: 0;
-    padding: 0;
-
-    li {
-      list-style: none;
-
-      &:not(:last-child) {
-        margin-bottom: 10px;
-      }
-
-      a {
-        color: var(--light);
-        text-decoration: none;
-        transition-duration: 0.3s;
-
-        &:hover {
-          color: var(--darkBrown);
-        }
-      }
-    }
-  }
-`;
-
-const Copyright = styled.div`
-  margin-top: 50px;
-  font-size: 0.85rem;
-  text-align: center;
-  color: var(--darkBrown);
-`;
+import * as Styles from "./styles";
 
 const Footer = () => (
-  <Wrapper>
-    <FooterMenu>
+  <Styles.Wrapper>
+    <Styles.FooterMenu>
       <div>
         <h4>Contactos</h4>
         <ul>
@@ -92,9 +45,11 @@ const Footer = () => (
           </li>
         </ul>
       </div>
-    </FooterMenu>
-    <Copyright>{["Copyright © ", new Date().getFullYear(), " Isto Vai Dar Merda. Toda a merda reservada."]}</Copyright>
-  </Wrapper>
+    </Styles.FooterMenu>
+    <Styles.Copyright>
+      {["Copyright © ", new Date().getFullYear(), " Isto Vai Dar Merda. Toda a merda reservada."]}
+    </Styles.Copyright>
+  </Styles.Wrapper>
 );
 
 export default Footer;
