+++ src/components/partners/types.d.js
@@ -1,0 +1,12 @@
+/**
+ * @typedef {object} IPartnerProps
+ * @property {string} [bg]
+ * @property {string} logo
+ * @property {string} name
+ * @property {string} url
+ */
+
+/**
+ * @typedef {object} IPartnersProps
+ * @property {IPartnerProps[]} partners
+ */
